+++ CHANGELOG.md
@@ -5,7 +5,28 @@ Releases
 ===============
 
 1. 24 Feb 2016 - 0.1.1-incubating
+2. 09 Mar 2015 - 16.03.01.RELEASE
 
+16.03.01.RELEASE
+=============
+
+This release includes following features and Bug fixes as mentioned below:
+ 
+Platform & API
+
+New Features & Improvements
+
+ - Added organization start date global configuration
+ - Center rescheduling
+ - Support for Interest Rate Charts based on amount
+  
+Bugs
+ 
+ - Infinite loop fix when retrying for loan schedular service
+ - Fix for repayment information
+ - [FINERACT-59] corrected interest calculation for flat loans
+ - [FINERACT-62] payment type applicable for disbursement charge configuration
+ 
 0.1.1-incubating
 =============